Remove and comment out old functionality
Some functionality has been commented out because it might need to be reimplemented in the TA in the near future.
