+++ ap_src/ap_app/absences.py
@@ -294,3 +294,59 @@ def add_recurring(request) -> render:
 
     content = {"form": form, "form2": form2}
     return render(request, "ap_app/add_recurring_absence.html", content)
+
+#JC - Add an absence via the form
+def manual_add(request:HttpRequest) -> render:
+    #JC - POST request
+    if request.method == "POST":
+        form = AbsenceForm(
+            request.POST, 
+            user=request.user,
+            initial={
+                "start_date": datetime.datetime.now(),
+                "end_date": lambda: datetime.datetime.now().date() + datetime.timedelta(days=1)
+            }
+        )
+        form.fields["user"].queryset = UserProfile.objects.filter(
+            edit_whitelist__in=[request.user]
+        )
+
+        #JC - Create absence
+        if form.is_valid():
+            absence = Absence()
+            absence.absence_date_start = form.cleaned_data["start_date"]
+            absence.absence_date_end = form.cleaned_data["end_date"]
+            absence.User_ID = request.user
+            absence.Target_User_ID = form.cleaned_data["user"].user
+
+            #JC - Check if the dates overlap with an existing absence. 
+            valid = True
+            Range = namedtuple('Range', ['start', 'end'])
+            r1 = Range(start=absence.absence_date_start, end=absence.absence_date_end)
+            for x in Absence.objects.filter(Target_User_ID=form.cleaned_data["user"].user.id):
+                r2 = Range(start=x.absence_date_start, end=x.absence_date_end)
+                delta = (min(r1.end, r2.end) -max(r1.start, r2.start)).days + 1
+                overlapp = max(0, delta)
+                if overlapp == 1:
+                    valid = False
+            
+            if valid:
+                absence.save()
+                return redirect("/")
+            else:
+                return render(request, "ap_app/add_absence.html", {
+                    "form": form,
+                    "message": "The absence conflicts with an existing absence",
+                    "message_type": "is-danger"
+                })
+
+    #JC - GET request
+    else:
+        form = AbsenceForm(user=request.user)
+        #JC - Allow users to edit others absence if they have permission
+        form.fields["user"].queryset = UserProfile.objects.filter(
+            edit_whitelist__in=[request.user]
+        )
+    
+    content = {"form": form}
+    return render(request, "ap_app/add_absence.html", content)
