+++ ap_src/ap_app/urls.py
@@ -17,20 +17,16 @@ urlpatterns = [
     path("calendar/<str:month>/<int:year>", views.main_calendar, name="all_calendar"),
     path("teams/", views.teams_dashboard, name="dashboard"),
     path("teams/create", views.create_team, name="create_team"),
-    path("teams/invite/", views.view_invites, name="view_invites"),
-    path("teams/invite/<int:team_id>/<int:user_id>/<str:role>", views.team_invite, name="team_invite"),
+    #path("teams/invite/", views.view_invites, name="view_invites"),
+    #path("teams/invite/<int:team_id>/<int:user_id>/<str:role>", views.team_invite, name="team_invite"),
     path("teams/join", views.join_team, name="join_team"),
-    path("teams/join/<int:id>/<str:role>", views.joining_team_process, name="join_team_2"),
-    path("teams/join/apply/<int:id>/<str:response>", views.joining_team_request, name="joining_team_request"),
-    path("teams/leave/<int:id>", views.leave_team, name="leave_team"),
-    path("teams/settings/<int:id>/", views.team_settings, name="team_settings"),
+    #path("teams/join/apply/<int:id>/<str:response>", views.joining_team_request, name="joining_team_request"),
     path("teams/api-calendar/<int:id>", views.api_team_calendar, name="api_team_calendar"),
     path("teams/api-calendar/<int:id>/<str:month>/<int:year>", views.api_team_calendar, name="api_team_calendar"),
-    path("teams/settings/<int:id>/<int:user_id>", views.edit_team_member_absence, name="edit_team_member_absence"),
-    path("teams/settings/promote/<int:id>/<int:user_id>", views.promote_team_member, name="promote_team_member"),
-    path("teams/settings/demote/<int:id>/<int:user_id>", views.demote_team_member, name="demote_team_member"),
-    path("teams/settings/remove/<int:id>/<int:user_id>", views.remove_team_member, name="remove_team_member"),
-    path("teams/misc/<int:id>", views.team_misc, name="misc"),
+    #path("teams/settings/promote/<int:id>/<int:user_id>", views.promote_team_member, name="promote_team_member"),
+    #path("teams/settings/demote/<int:id>/<int:user_id>", views.demote_team_member, name="demote_team_member"),
+    #path("teams/settings/remove/<int:id>/<int:user_id>", views.remove_team_member, name="remove_team_member"),
+    #path("teams/misc/<int:id>", views.team_misc, name="misc"),
     path("teams/edit/<int:id>", views.edit_team, name="edit_team"),
     path("absence/add", teams.manual_add, name="add"),
     path("absence/add_recurring", views.add_recurring, name="add_recurring"),
